Move ProjectCards and other common to lib
Moved ProjectCard, ProjectFilterSelector, and the two common.ts scripts below `routes/projects` to `lib/projects`
