+++ source/SIL.AppBuilder.Portal/src/auth.ts
@@ -4,7 +4,7 @@ import Auth0Provider from '@auth/sveltekit/providers/auth0';
 import { error, redirect, type Handle } from '@sveltejs/kit';
 import { DatabaseWrites, prisma } from 'sil.appbuilder.portal.common';
 import { RoleId } from 'sil.appbuilder.portal.common/prisma';
-import { verifyCanViewAndEdit } from './routes/(authenticated)/projects/[id=idNumber]/common';
+import { verifyCanViewAndEdit } from './lib/projects/common.server';
 
 declare module '@auth/sveltekit' {
   interface Session {
