+++ source/SIL.AppBuilder.Portal/src/lib/projects/common.ts
