+++ source/SIL.AppBuilder.Portal/src/lib/projects/components/ProjectCard.svelte
